pyproject.toml
[tool.poetry]
name = "dacutil"
version = "0.3.3"
description = "For Data Analytic"
authors = ["Attapon Thanawong <attapon_srem@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"

pandas = ">=2.1.0"
pyarrow = ">=14.0.0"
requests = ">=2.31.0"
configobj = ">=5.0.0"
pyyaml = ">=6.0.0"
tomli = ">=2.0.0"
cryptography = ">=41.0.0"
pyrage = ">=1.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
